fix: viewCnt int 타입에서 long 타입으로 변경 및 likeCnt heartCnt로 변수명 변경

+++ src/main/java/com/gaethering/gaetheringserver/domain/board/dto/PostWriteResponse.java
@@ -22,9 +22,9 @@ public class PostWriteResponse {
 
     private String categoryName;
 
-    private int viewCnt;
+    private long viewCnt;
 
-    private int heartCnt;
+    private long heartCnt;
 
     private String nickname;
